move multisig plugins model
