+++ src/interfaces/index.ts
@@ -2,5 +2,5 @@ export * from './account'
 export * from './chain'
 export * from './createAccount'
 export * from './transaction'
-export * from './plugins/multisigTransaction'
-export * from './plugins/multisigCreateAccount'
+export * from './plugins/multisig/multisigPluginTransaction'
+export * from './plugins/multisig/multisigPluginCreateAccount'
