+++ src/interfaces/plugins/multisig/index.ts
@@ -1,0 +1,3 @@
+export * from './multisigPlugin'
+export * from './multisigPluginCreateAccount'
+export * from './multisigPluginTransaction'
